Changed the folder structure of the unit tests
