--- tests/unit_ops/test_charm.py
@@ -1,134 +1,0 @@
-# Copyright 2025 Canonical Ltd.
-# See LICENSE file for licensing details.
-
-"""Discourse K8s operator charm unit tests."""
-
-import pytest
-from ops import testing
-from ops.model import ActiveStatus, BlockedStatus
-
-from charm import CONTAINER_NAME, INVALID_CORS_MESSAGE, SERVICE_NAME, DiscourseCharm
-
-
-@pytest.mark.parametrize(
-    "config, expected_origin, expected_status",
-    [
-        pytest.param(
-            {
-                "cors_origin": "*",
-                "augment_cors_origin": True,
-                "external_hostname": "example.com",
-                "force_https": True,
-                "s3_cdn_url": "https://cdn.test",
-            },
-            "*",
-            ActiveStatus(),
-            id="Wildcard disables augmentation",
-        ),
-        pytest.param(
-            {
-                "cors_origin": "",
-                "augment_cors_origin": False,
-                "external_hostname": "example.com",
-                "force_https": True,
-                "s3_cdn_url": "https://cdn.test",
-            },
-            "*",
-            BlockedStatus(INVALID_CORS_MESSAGE),
-            id="Raise error when invalid CORS config",
-        ),
-        pytest.param(
-            {
-                "cors_origin": "",
-                "augment_cors_origin": True,
-                "external_hostname": "example.com",
-                "force_https": True,
-                "s3_cdn_url": "",
-            },
-            "https://example.com",
-            ActiveStatus(),
-            id="Augment only with external_hostname (HTTPS)",
-        ),
-        pytest.param(
-            {
-                "cors_origin": "",
-                "augment_cors_origin": True,
-                "external_hostname": "example.com",
-                "force_https": False,
-                "s3_cdn_url": "https://cdn.test",
-            },
-            "http://example.com,https://cdn.test",
-            ActiveStatus(),
-            id="Augment with both external_hostname (HTTP) and s3_cdn_url",
-        ),
-        pytest.param(
-            {
-                "cors_origin": "https://custom.origin",
-                "augment_cors_origin": False,
-                "external_hostname": "example.com",
-                "force_https": True,
-                "s3_cdn_url": "https://cdn.test",
-            },
-            "https://custom.origin",
-            ActiveStatus(),
-            id="User-defined cors_origin, no augmentation",
-        ),
-        pytest.param(
-            {
-                "cors_origin": "https://custom.origin",
-                "augment_cors_origin": True,
-                "external_hostname": "example.com",
-                "force_https": True,
-                "s3_cdn_url": "https://cdn.test",
-            },
-            "https://cdn.test,https://custom.origin,https://example.com",
-            ActiveStatus(),
-            id="User-defined cors_origin with augmentation enabled",
-        ),
-        pytest.param(
-            {
-                "cors_origin": "https://foo.com, https://bar.com",
-                "augment_cors_origin": True,
-                "external_hostname": "example.com",
-                "force_https": False,
-                "s3_cdn_url": "https://cdn.test",
-            },
-            "http://example.com,https://bar.com,https://cdn.test,https://foo.com",
-            ActiveStatus(),
-            id="Multiple user-defined cors_origins with augmentation",
-        ),
-        pytest.param(
-            {
-                "cors_origin": " https://foo.com , https://foo.com ",
-                "augment_cors_origin": True,
-                "external_hostname": "foo.com",
-                "force_https": True,
-                "s3_cdn_url": "https://foo.com",
-            },
-            "https://foo.com",
-            ActiveStatus(),
-            id="Duplicated origins across cors_origin and augmentation",
-        ),
-    ],
-)
-def test_get_cors_origin_behavior(config, expected_origin, expected_status, base_state):
-    """
-    arrange: deploy charm with CORS-related config
-    act: configure charm with varying CORS inputs
-    assert: DISCOURSE_CORS_ORIGIN matches expected result
-    """
-    ctx = testing.Context(DiscourseCharm)
-
-    base_state["config"] = config
-
-    state_in = testing.State(**base_state)
-    container = state_in.get_container(CONTAINER_NAME)
-
-    state_out = ctx.run(ctx.on.pebble_ready(container), state_in)
-    plan_out = state_out.get_container(container.name).plan
-
-    assert state_out.unit_status == expected_status
-    if expected_status == ActiveStatus():  # plan is empty when in BlockedStatus
-        assert (
-            plan_out.services[SERVICE_NAME].environment["DISCOURSE_CORS_ORIGIN"] == expected_origin
-        )
