+++ tox.ini
@@ -120,7 +120,7 @@ deps =
     git+https://github.com/canonical/saml-test-idp.git
     -r{toxinidir}/requirements.txt
 commands =
-    pytest -v --tb native --ignore={[vars]tst_path}unit --ignore={[vars]tst_path}unit_ops --log-cli-level=INFO -s {posargs}
+    pytest -v --tb native --ignore={[vars]tst_path}unit --ignore={[vars]tst_path}unit_harness --log-cli-level=INFO -s {posargs}
 
 [testenv:src-docs]
 allowlist_externals=sh
